Added rain powerup and the ability for the runner to grow and shrink
Added the ability to make it rain and for the rain to make the runner grow. After a certain amount of time, the runner will shrink back to their initial size. Also gave the runner gravity and collision with the ground. If they are in the air, they will fall back to the ground.